# apps/backend/openapi.yaml
openapi: 3.1.0
info:
  title: AI Voice Assistant & Calendar Manager API
  description: |
    FastAPI backend for the AI-powered voice assistant and calendar manager.
    
    ## Features
    - Voice processing and natural language understanding
    - Calendar management and event scheduling
    - AI integration with OpenAI GPT-4 and Anthropic Claude
    - Real-time WebSocket communication
    - User authentication and authorization
    
    ## Authentication
    This API uses JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - 100 requests per minute for authenticated users
    - 10 requests per minute for unauthenticated users
    
  version: 1.0.0
  contact:
    name: AI Voice Assistant Team
    email: support@aivoiceassistant.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.aivoiceassistant.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/v1/calendar/events:
    get:
      summary: Get Events
      description: Retrieve calendar events for the authenticated user
      tags:
        - Calendar
      security:
        - BearerAuth: []
      parameters:
        - name: start_date
          in: query
          description: Start date for event range (ISO 8601)
          required: false
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          description: End date for event range (ISO 8601)
          required: false
          schema:
            type: string
            format: date-time
        - name: calendar_id
          in: query
          description: Filter by specific calendar ID
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Events retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create Event
      description: Create a new calendar event
      tags:
        - Calendar
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Event'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/calendar/events/{event_id}:
    get:
      summary: Get Event
      description: Retrieve a specific calendar event
      tags:
        - Calendar
      security:
        - BearerAuth: []
      parameters:
        - name: event_id
          in: path
          required: true
          description: Event ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Event retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update Event
      description: Update an existing calendar event
      tags:
        - Calendar
      security:
        - BearerAuth: []
      parameters:
        - name: event_id
          in: path
          required: true
          description: Event ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventRequest'
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete Event
      description: Delete a calendar event
      tags:
        - Calendar
      security:
        - BearerAuth: []
      parameters:
        - name: event_id
          in: path
          required: true
          description: Event ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Event deleted successfully
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Event's unique identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        title:
          type: string
          description: Event title
          example: "Team Meeting"
        description:
          type: string
          description: Event description
          example: "Weekly team standup meeting"
        start_time:
          type: string
          format: date-time
          description: Event start time
          example: "2024-01-15T09:00:00Z"
        end_time:
          type: string
          format: date-time
          description: Event end time
          example: "2024-01-15T10:00:00Z"
        location:
          type: string
          description: Event location
          example: "Conference Room A"
        calendar_id:
          type: string
          format: uuid
          description: Associated calendar ID
          example: "123e4567-e89b-12d3-a456-426614174000"
        user_id:
          type: string
          format: uuid
          description: Event owner's user ID
          example: "123e4567-e89b-12d3-a456-426614174000"
        attendees:
          type: array
          items:
            $ref: '#/components/schemas/Attendee'
        reminders:
          type: array
          items:
            $ref: '#/components/schemas/Reminder'
        status:
          type: string
          enum: [confirmed, tentative, cancelled]
          description: Event status
          example: "confirmed"
        priority:
          type: string
          enum: [low, medium, high, urgent]
          description: Event priority
          example: "medium"
        tags:
          type: array
          items:
            type: string
          description: Event tags
          example: ["work", "meeting"]
        created_at:
          type: string
          format: date-time
          description: Event creation timestamp
          example: "2024-01-10T08:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Event last update timestamp
          example: "2024-01-15T12:00:00Z"
    CreateEventRequest:
      type: object
      required:
        - title
        - start_time
        - end_time
      properties:
        title:
          type: string
          minLength: 1
          description: Event title
          example: "Team Meeting"
        description:
          type: string
          description: Event description
          example: "Weekly team standup meeting"
        start_time:
          type: string
          format: date-time
          description: Event start time
          example: "2024-01-15T09:00:00Z"
        end_time:
          type: string
          format: date-time
          description: Event end time
          example: "2024-01-15T10:00:00Z"
        location:
          type: string
          description: Event location
          example: "Conference Room A"
        calendar_id:
          type: string
          format: uuid
          description: Associated calendar ID
          example: "123e4567-e89b-12d3-a456-426614174000"
        attendees:
          type: array
          items:
            type: string
            format: email
          description: List of attendee email addresses
          example: ["user1@example.com", "user2@example.com"]
        reminders:
          type: array
          items:
            $ref: '#/components/schemas/ReminderRequest'
        status:
          type: string
          enum: [confirmed, tentative, cancelled]
          description: Event status
          example: "confirmed"
        priority:
          type: string
          enum: [low, medium, high, urgent]
          description: Event priority
          example: "medium"
        tags:
          type: array
          items:
            type: string
          description: Event tags
          example: ["work", "meeting"]
    UpdateEventRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          description: Event title
          example: "Team Meeting"
        description:
          type: string
          description: Event description
          example: "Weekly team standup meeting"
        start_time:
          type: string
          format: date-time
          description: Event start time
          example: "2024-01-15T09:00:00Z"
        end_time:
          type: string
          format: date-time
          description: Event end time
          example: "2024-01-15T10:00:00Z"
        location:
          type: string
          description: Event location
          example: "Conference Room A"
        attendees:
          type: array
          items:
            type: string
            format: email
          description: List of attendee email addresses
          example: ["user1@example.com", "user2@example.com"]
        reminders:
          type: array
          items:
            $ref: '#/components/schemas/ReminderRequest'
        status:
          type: string
          enum: [confirmed, tentative, cancelled]
          description: Event status
          example: "confirmed"
        priority:
          type: string
          enum: [low, medium, high, urgent]
          description: Event priority
          example: "medium"
        tags:
          type: array
          items:
            type: string
          description: Event tags
          example: ["work", "meeting"]
    Attendee:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Attendee's unique identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        email:
          type: string
          format: email
          description: Attendee's email address
          example: "user@example.com"
        name:
          type: string
          description: Attendee's name
          example: "John Doe"
        response:
          type: string
          enum: [accepted, declined, pending, tentative]
          description: Attendee's response to the event
          example: "accepted"
    Reminder:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Reminder's unique identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        event_id:
          type: string
          format: uuid
          description: Associated event ID
          example: "123e4567-e89b-12d3-a456-426614174000"
        time:
          type: string
          format: date-time
          description: Reminder time
          example: "2024-01-15T08:45:00Z"
        type:
          type: string
          enum: [email, push, sms, voice]
          description: Reminder type
          example: "push"
        sent:
          type: boolean
          description: Whether the reminder has been sent
          example: false
    ReminderRequest:
      type: object
      required:
        - time
        - type
      properties:
        time:
          type: string
          format: date-time
          description: Reminder time
          example: "2024-01-15T08:45:00Z"
        type:
          type: string
          enum: [email, push, sms, voice]
          description: Reminder type
          example: "push"
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
              example: "validation_error"
            message:
              type: string
              description: Error message
              example: "Invalid input data"
            details:
              type: object
              description: Additional error details
              example:
                field: "email"
                issue: "Invalid email format"
...